refactor: HttpStatusCodes into application/types
