--- src/apps/backend/modules/application/common/http/index.ts
@@ -1,1 +1,0 @@
-export * from './status-codes';
