--- src/apps/backend/modules/application/common/http/status-codes.ts
@@ -1,13 +1,0 @@
-export enum HttpStatusCodes {
-  OK = 200,
-  CREATED = 201,
-  NO_CONTENT = 204,
-  BAD_REQUEST = 400,
-  UNAUTHORIZED = 401,
-  NOT_FOUND = 404,
-  CONFLICT = 409,
-  SERVER_ERROR = 500,
-  SERVICE_UNAVAILABLE = 503,
-}
-
-export default HttpStatusCodes;
