+++ src/apps/backend/modules/application/index.ts
@@ -3,3 +3,4 @@ export { default as ApplicationRepository } from './application-repository';
 export { default as ApplicationRouter } from './application-router';
 export { default as ApplicationServer } from './application-server';
 export * from './application-controller';
+export * from './types';
