+++ src/apps/backend/modules/communication/types.ts
@@ -1,5 +1,4 @@
-import { ApplicationError } from '../application';
-import { HttpStatusCodes } from '../application/common/http';
+import { ApplicationError, HttpStatusCodes } from '../application';
 
 export interface LooseObject {
   [key: string]: unknown;
