+++ src/apps/backend/modules/documentation/types.ts
@@ -1,5 +1,4 @@
-import { ApplicationError } from '../application';
-import { HttpStatusCodes } from '../application/common/http';
+import { ApplicationError, HttpStatusCodes } from '../application';
 
 import { HttpRoute } from './list-routes';
