+++ src/apps/backend/modules/password-reset-token/rest-api/password-reset-token-controller.ts
@@ -1,5 +1,4 @@
-import { applicationController, Request, Response } from '../../application';
-import { HttpStatusCodes } from '../../application/common/http';
+import { applicationController, HttpStatusCodes, Request, Response } from '../../application';
 import PasswordResetTokenService from '../password-reset-token-service';
 import { CreatePasswordResetTokenParams } from '../types';
