+++ src/apps/backend/modules/password-reset-token/types.ts
@@ -1,5 +1,4 @@
-import { ApplicationError } from '../application';
-import { HttpStatusCodes } from '../application/common/http';
+import { ApplicationError, HttpStatusCodes } from '../application';
 
 export class PasswordResetToken {
   id: string;
