+++ src/apps/backend/modules/task/rest-api/task-controller.ts
@@ -1,5 +1,4 @@
-import { applicationController, Request, Response } from '../../application';
-import { HttpStatusCodes } from '../../application/common/http';
+import { applicationController, HttpStatusCodes, Request, Response } from '../../application';
 import TaskService from '../task-service';
 import {
   Task,
